refactor and consolidate functions
