+++ miner.py
@@ -12,8 +12,6 @@ import base64
 import requests
 import time
 
-from filelock import FileLock
-
 from webcash import (
     SecretWebcash,
     compute_target,
@@ -25,26 +23,11 @@ from walletclient import (
     create_webcash_wallet,
 )
 
+from utils import lock_wallet
+
 INTERVAL_LENGTH_IN_SECONDS = 10
 
 WALLET_FILENAME = "default_wallet.webcash"
-
-# Don't want to name this anything close to the wallet filename because the
-# user may have to manually delete the lock at some point.
-LOCKFILE_NAME = "LOCK"
-
-# disable the filelock timeout by setting a negative value
-lock = FileLock(LOCKFILE_NAME, timeout=-1)
-
-def lock_wallet(func):
-    """
-    Function decorator for locking the file.
-    """
-    def wrapper(*args, **kwargs):
-        with lock:
-            result = func(*args, **kwargs)
-        return result
-    return wrapper
 
 def get_protocol_settings():
     response = requests.get("https://webcash.tech/api/v1/target")
