+++ walletclient.py
@@ -24,7 +24,6 @@ import sys
 
 import requests
 import click
-from filelock import FileLock
 
 #from miner import mine
 from webcash import (
@@ -35,27 +34,12 @@ from webcash import (
     deserialize_amount,
 )
 
+from utils import lock_wallet
+
 # unused?
 FEE_AMOUNT = 0
 
 WALLET_NAME = "default_wallet.webcash"
-
-# Don't want to name this anything close to the wallet filename because the
-# user may have to manually delete the lock at some point.
-LOCKFILE_NAME = "LOCK"
-
-# disable the filelock timeout by setting a negative value
-lock = FileLock(LOCKFILE_NAME, timeout=-1)
-
-def lock_wallet(func):
-    """
-    Function decorator for locking the file.
-    """
-    def wrapper(*args, **kwargs):
-        with lock:
-            result = func(*args, **kwargs)
-        return result
-    return wrapper
 
 # TODO: decryption
 def load_webcash_wallet(filename=WALLET_NAME):
